requirements.txt
Pillow
PyPDF2
tinytag
olefile
GPSPhoto
exifread
piexif